+++ apps/cookbook/src/app/examples/slides-example/examples/_shared.scss
@@ -11,3 +11,13 @@
 kirby-card {
   justify-content: start;
 }
+
+.kirby-text-medium,
+.kirby-text-medium h1,
+.kirby-text-medium h2,
+.kirby-text-medium h3,
+.kirby-text-medium h4,
+.kirby-text-medium h5,
+.kirby-text-medium h6 {
+  margin: 0;
+}
