+++ apps/cookbook/src/app/examples/slides-example/examples/advanced.ts
