+++ apps/cookbook/src/app/examples/slides-example/examples/custom-heading.ts
@@ -3,8 +3,9 @@ import { SlideModule } from '@kirbydesign/designsystem/slide';
 import { CardModule } from '@kirbydesign/designsystem/card';
 
 const config = {
-  selector: 'cookbook-slides-simple-example',
-  template: `<kirby-slides [slides]="slides" [title]="'Title'" [showNavigation]="true">
+  selector: 'cookbook-slides-custom-heading-example',
+  template: `<kirby-slides [slides]="slides"   [title]="'Title'" [showNavigation]="true">    
+  <h4 class="kirby-text-medium">Custom heading</h4>
   <kirby-card *kirbySlide="let slide; let i = index" slideStretchHeight [hasPadding]="true">
     <kirby-card-header [title]="slide.title" [subtitle]="slide.subtitle"></kirby-card-header>
     <div class="card-content">
@@ -20,9 +21,8 @@ const config = {
   template: config.template,
   imports: [SlideModule, CardModule],
 })
-export class SlidesSimpleExampleComponent {
+export class SlidesCustomHeadingExampleComponent {
   template = config.template;
-  title = `Title`;
   slides = [...Array(9).keys()].map((number) => ({
     title: `Slide ${number + 1}`,
     subtitle: `Subtitle ${number + 1}`,
