+++ apps/cookbook/src/app/examples/slides-example/slides-example.component.html
@@ -7,15 +7,15 @@
     <cookbook-slides-simple-example></cookbook-slides-simple-example>
   </div>
 
-  <h2>Slides custom heading example</h2>
-
-  <div class="example-frame">
-    <cookbook-slides-custom-heading-example></cookbook-slides-custom-heading-example>
-  </div>
-
   <h2>Slides height example</h2>
 
   <div class="example-frame">
     <cookbook-slides-height-example></cookbook-slides-height-example>
+  </div>
+
+  <h2>Slides custom heading example</h2>
+
+  <div class="example-frame">
+    <cookbook-slides-custom-heading-example></cookbook-slides-custom-heading-example>
   </div>
 </div>
