+++ apps/cookbook/src/app/showcase/_showcase.shared.scss
@@ -18,7 +18,6 @@ cookbook-code-viewer {
   border: 1px solid #ddd;
   border-radius: 12px;
   padding: utils.size('m');
-  margin-bottom: 24px;
 
   &.no-padding {
     padding: 0;
