+++ apps/cookbook/src/app/showcase/slides-showcase/slides-showcase.component.html
@@ -39,9 +39,16 @@
   </div>
 </cookbook-example-viewer>
 
-<h2>Custom heading</h2>
+<h2>Heading level</h2>
 
-<p>If you need a custom heading for the slides, you can add h1,h2,h3,h4,h5 or h6 tag.</p>
+<p>
+  If you need a custom heading for the slides, you can add
+  <code>&lt;h1&gt;</code>
+  –
+  <code>&lt;h6&gt;</code>
+  . This allows slide titles to become part of the page’s heading structure, making them accessible
+  to screen readers.
+</p>
 
 <cookbook-example-viewer [html]="customHeadingExample.template">
   <div class="example-frame">
